Initial Seed Script Initial Seed Script
**Why**
We want the ability to seed specific tables with data as the database boots up. This will allow us to easily populate tables with values.

**How**
Create a seeder interface and seed file for each table we wish to seed. Add runSeed function to main that will add all seed functions and run them. Update migrations to handle the ability to remove certain columns if they exist. Clean up penaltyType table.