[issue#78/issue#79] Pagination + Sorting POC (#22)
* [issue#78/issue#79] Pagination + Sorting POC
[issue#78/issue#79] Pagination + Sorting POC

**Why**
We want to be able to paginate and sort our backend endpoint to allow efficient queries to the database as the number of records grow.

**How**
Add pagination and sorting to the league end point. Created new field validator that can check if the passed in field is a valid field. Updated yml files due to errors on schema imports.

* Update all OpenAPISpecs to have correct references. Update OpenAPISpecs to use component sections as required by documentation. Make sort and pagination params re-usable. Update tests and validators to support more generic usage.

* Fix error in accidentally deleted text. Remove comments.

* Fix indentation issue.

* Remove todos.

* Removed unreferenced yaml.