Standardize Models Standardize Models
**Why**
We want to standardize the models to better reflect what they actually are. The main aim is to rename League -> LeagueRecord as it is an instance of a League to a specific season. The pattern would follow for Teams if we decide to go this route.

**How**
Rename League -> LeagueRecord. Rename CorrelationId -> LeagueID.